Move TreeNodeBusy
