+++ bundles/map-editor/src/main/java/org/orbisgis/mapeditor/map/mapsManager/TreeNodeMapCatalogServer.java
@@ -41,6 +41,8 @@ import javax.swing.ImageIcon;
 import javax.swing.JMenuItem;
 import javax.swing.JPopupMenu;
 import javax.swing.tree.MutableTreeNode;
+
+import org.orbisgis.sif.components.resourceTree.TreeNodeBusy;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.orbisgis.corejdbc.DataManager;
