+++ bundles/map-editor/src/main/java/org/orbisgis/mapeditor/map/mapsManager/TreeNodeWorkspace.java
@@ -40,6 +40,8 @@ import javax.swing.JMenuItem;
 import javax.swing.JPopupMenu;
 import javax.swing.TransferHandler.TransferSupport;
 import javax.swing.tree.MutableTreeNode;
+
+import org.orbisgis.sif.components.resourceTree.TreeNodeBusy;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.orbisgis.coremap.layerModel.MapContext;
