+++ orbisgis-sif/src/main/java/org/orbisgis/sif/components/fstree/TreeNodeBusy.java
@@ -26,15 +26,14 @@
  * or contact directly:
  * info_at_ orbisgis.org
  */
-package org.orbisgis.mapeditor.map.mapsManager;
+package org.orbisgis.sif.components.fstree;
 
 import java.awt.event.ActionListener;
 import java.beans.EventHandler;
 import javax.swing.ImageIcon;
 import javax.swing.Timer;
-import org.orbisgis.mapeditor.map.icons.MapEditorIcons;
-import org.orbisgis.sif.components.fstree.AbstractTreeNodeLeaf;
-import org.orbisgis.sif.components.fstree.TreeNodeCustomIcon;
+
+import org.orbisgis.sif.icons.SifIcon;
 import org.xnap.commons.i18n.I18n;
 import org.xnap.commons.i18n.I18nFactory;
 
@@ -103,7 +102,7 @@ public class TreeNodeBusy extends AbstractTreeNodeLeaf implements TreeNodeCustom
 
         @Override
         public ImageIcon getLeafIcon() {
-                return MapEditorIcons.getIcon(frames[curFrame]);
+                return SifIcon.getIcon(frames[curFrame]);
         }
 
         @Override
