+++ orbisgis-sif/src/main/resources/org/orbisgis/sif/icons/progress_4.png
